chore: move contracts into contracts/
