+++ src/contracts/erc20.js
@@ -1,4 +1,4 @@
-import Contract from './contract';
+import Contract from '../contract';
 
 const ERC20_ABI = [
   {
